added name of app into param

+++ src/hooks/useConnectApp.js
@@ -2,12 +2,16 @@ import UserService from '../services/User.service'
 export const useConnectAppMixin = {
   methods: {
     async connectApp () {
-      window.open(process.env.VUE_APP_ALPACA_INTEGRATION_URL, '_self')
+      if (name === 'Alpaca') {
+        window.open(process.env.VUE_APP_ALPACA_INTEGRATION_URL, '_self')
+      }
     },
-    async removeApp (userId, store) {
-      await UserService.removeAlpaca(userId, store)
-      this.link = false
-      this.$emit('linked', false)
+    async removeApp (userId, store, name) {
+      if (name === 'Alpaca') {
+        await UserService.removeAlpaca(userId, store)
+        this.link = false
+        this.$emit('linked', false)
+      }
     }
   }
 }
